vfp: force non-conditional encoding for external Thumb2 tail call
Nick reports that the following error is produced in some cases when
using GCC+ld.bfd to build the ARM defconfig with Thumb2 enabled:

  arch/arm/vfp/vfphw.o: in function `vfp_support_entry':
  (.text+0xa): relocation truncated to fit: R_ARM_THM_JUMP19 against
  symbol `vfp_kmode_exception' defined in .text.unlikely section in
  arch/arm/vfp/vfpmodule.o

  $ arm-linux-gnueabihf-ld --version
  GNU ld (GNU Binutils for Debian) 2.34

Generally, the linker should be able to fix up out of range branches by
emitting veneers, but apparently, it fails to do so in this particular
case, i.e., a conditional 'tail call' to vfp_kmode_exception(), which
is not defined in the same object file.

So let's force the use of a non-conditional encoding of the B instruction,
which has more space for an immediate offset.

Cc: <(address hidden)>
Reported-by: Nick Desaulniers <(address hidden)>
Tested-by: Nick Desaulniers <(address hidden)>
Signed-off-by: Ard Biesheuvel <(address hidden)>
Signed-off-by: Aníbal Limón <limon.anibal@gmail.com>

From: https://www.armlinux.org.uk/developer/patches/viewpatch.php?id=9018/1